Initial commit of refactored frontend (squashed commit)
New frontend is using Angular instead of AngularJS. Number of files
exploded - frontend itself contains more files and weights more than
other components of projects combined.

spec.ts files in this commit are templates generated by Angular-cli and
are not meant to be run. Checks are kept in separate branch for now.

Of course there is still no documentation.